requirements.txt
customtkinter
requests
pydub
unidecode
pygame
CTkMessagebox
CTkToolTip
num2words
ffmpeg-python
pdftextract
regex
hasami
ebooklib
beautifulsoup4
mutagen
Pillow
argparse
sentence-splitter
yt_dlp
gtts
flask
pydub
